requirements.txt
# python-telegram-bot>=22.5
telethon>=1.42.0
python-dotenv>=1.2.1
pyyaml>=6.0.3
SQLAlchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0